requirement.txt
pandas==1.2.3
python-docx==0.8.10
pdfminer.six==20201018 
requests==2.25.1
Flask==1.1.2
setuptools==56.0.0
Pillow==8.2.0
lxml==4.6.3